+++ packages/core/src/routes/init.ts
@@ -10,9 +10,9 @@ import applicationRoutes from '@/routes/application';
 import connectorRoutes from '@/routes/connector';
 import dashboardRoutes from '@/routes/dashboard';
 import resourceRoutes from '@/routes/resource';
+import sessionPasswordlessRoutes from '@/routes/session/passwordless';
 import sessionRoutes from '@/routes/session/session';
-import sessionPasswordlessRoutes from '@/routes/session/session-passwordless';
-import sessionSocialRoutes from '@/routes/session/session-social';
+import sessionSocialRoutes from '@/routes/session/social';
 import settingRoutes from '@/routes/setting';
 import signInExperiencesRoutes from '@/routes/sign-in-experience';
 import signInSettingsRoutes from '@/routes/sign-in-settings';
